+++ pkg/logqlmodel/stats/context_test.go
@@ -25,7 +25,7 @@ func TestResult(t *testing.T) {
 	fakeIngesterQuery(ctx)
 	fakeIngesterQuery(ctx)
 
-	res := stats.Result(2 * time.Second)
+	res := stats.Result(2*time.Second, 2*time.Nanosecond)
 	res.Log(util_log.Logger)
 	expected := Result{
 		Ingester: Ingester{
@@ -61,6 +61,7 @@ func TestResult(t *testing.T) {
 		},
 		Summary: Summary{
 			ExecTime:                2 * time.Second.Seconds(),
+			QueueTime:               2 * time.Nanosecond.Seconds(),
 			BytesProcessedPerSecond: int64(42),
 			LinesProcessedPerSecond: int64(50),
 			TotalBytesProcessed:     int64(84),
@@ -106,6 +107,7 @@ func TestSnapshot_JoinResults(t *testing.T) {
 		},
 		Summary: Summary{
 			ExecTime:                2 * time.Second.Seconds(),
+			QueueTime:               2 * time.Nanosecond.Seconds(),
 			BytesProcessedPerSecond: int64(42),
 			LinesProcessedPerSecond: int64(50),
 			TotalBytesProcessed:     int64(84),
@@ -114,7 +116,7 @@ func TestSnapshot_JoinResults(t *testing.T) {
 	}
 
 	JoinResults(ctx, expected)
-	res := statsCtx.Result(2 * time.Second)
+	res := statsCtx.Result(2*time.Second, 2*time.Nanosecond)
 	require.Equal(t, expected, res)
 }
 
@@ -177,6 +179,7 @@ func TestResult_Merge(t *testing.T) {
 		},
 		Summary: Summary{
 			ExecTime:                2 * time.Second.Seconds(),
+			QueueTime:               2 * time.Nanosecond.Seconds(),
 			BytesProcessedPerSecond: int64(42),
 			LinesProcessedPerSecond: int64(50),
 			TotalBytesProcessed:     int64(84),
@@ -223,6 +226,7 @@ func TestResult_Merge(t *testing.T) {
 		},
 		Summary: Summary{
 			ExecTime:                2 * 2 * time.Second.Seconds(),
+			QueueTime:               2 * 2 * time.Nanosecond.Seconds(),
 			BytesProcessedPerSecond: int64(42), // 2 requests at the same pace should give the same bytes/lines per sec
 			LinesProcessedPerSecond: int64(50),
 			TotalBytesProcessed:     2 * int64(84),
@@ -234,10 +238,10 @@ func TestResult_Merge(t *testing.T) {
 func TestReset(t *testing.T) {
 	statsCtx, ctx := NewContext(context.Background())
 	fakeIngesterQuery(ctx)
-	res := statsCtx.Result(2 * time.Second)
+	res := statsCtx.Result(2*time.Second, 2*time.Millisecond)
 	require.NotEmpty(t, res)
 	statsCtx.Reset()
-	res = statsCtx.Result(0)
+	res = statsCtx.Result(0, 0)
 	require.Empty(t, res)
 }