+++ pkg/querier/queryrange/codec_test.go
@@ -903,10 +903,11 @@ var (
 		},
 		"summary": {
 			"bytesProcessedPerSecond": 20,
-			"execTime": 21,
-			"linesProcessedPerSecond": 22,
-			"totalBytesProcessed": 23,
-			"totalLinesProcessed": 24
+			"execTime": 22,
+			"linesProcessedPerSecond": 23,
+			"queueTime": 21,
+			"totalBytesProcessed": 24,
+			"totalLinesProcessed": 25
 		}
 	},`
 	matrixString = `{
@@ -1052,10 +1053,11 @@ var (
 	statsResult = stats.Result{
 		Summary: stats.Summary{
 			BytesProcessedPerSecond: 20,
-			ExecTime:                21,
-			LinesProcessedPerSecond: 22,
-			TotalBytesProcessed:     23,
-			TotalLinesProcessed:     24,
+			QueueTime:               21,
+			ExecTime:                22,
+			LinesProcessedPerSecond: 23,
+			TotalBytesProcessed:     24,
+			TotalLinesProcessed:     25,
 		},
 		Querier: stats.Querier{
 			Store: stats.Store{