+++ pkg/storage/chunk/openstack/swift_object_client.go
@@ -24,7 +24,8 @@ import (
 
 var defaultTransport http.RoundTripper = &http.Transport{
 	Proxy:                 http.ProxyFromEnvironment,
-	MaxIdleConnsPerHost:   512,
+	MaxIdleConnsPerHost:   200,
+	MaxIdleConns:          200,
 	ExpectContinueTimeout: 5 * time.Second,
 }
 
@@ -111,7 +112,11 @@ func createConnection(cfg SwiftConfig, hedgingCfg hedging.Config, hedging bool) 
 		c.DomainId = cfg.UserDomainID
 	}
 	if hedging {
-		c.Transport = hedgingCfg.RoundTripperWithRegisterer(c.Transport, prometheus.WrapRegistererWithPrefix("loki", prometheus.DefaultRegisterer))
+		var err error
+		c.Transport, err = hedgingCfg.RoundTripperWithRegisterer(c.Transport, prometheus.WrapRegistererWithPrefix("loki", prometheus.DefaultRegisterer))
+		if err != nil {
+			return nil, err
+		}
 	}
 
 	err := c.Authenticate()
@@ -127,15 +132,15 @@ func (s *SwiftObjectClient) Stop() {
 	s.hedgingConn.UnAuthenticate()
 }
 
-// GetObject returns a reader for the specified object key from the configured swift container.
-func (s *SwiftObjectClient) GetObject(ctx context.Context, objectKey string) (io.ReadCloser, error) {
+// GetObject returns a reader and the size for the specified object key from the configured swift container.
+func (s *SwiftObjectClient) GetObject(ctx context.Context, objectKey string) (io.ReadCloser, int64, error) {
 	var buf bytes.Buffer
 	_, err := s.hedgingConn.ObjectGet(s.cfg.ContainerName, objectKey, &buf, false, nil)
 	if err != nil {
-		return nil, err
+		return nil, 0, err
 	}
 
-	return ioutil.NopCloser(&buf), nil
+	return ioutil.NopCloser(&buf), int64(buf.Len()), nil
 }
 
 // PutObject puts the specified bytes into the configured Swift container at the provided key