+++ pkg/util/httpreq/tags_test.go
@@ -4,6 +4,7 @@ import (
 	"net/http"
 	"net/http/httptest"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -40,6 +41,47 @@ func TestQueryTags(t *testing.T) {
 			checked := false
 			mware := ExtractQueryTagsMiddleware().Wrap(http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
 				require.Equal(t, tc.exp, req.Context().Value(QueryTagsHTTPHeader).(string))
+				checked = true
+			}))
+
+			mware.ServeHTTP(w, req)
+
+			assert.True(t, true, checked)
+		})
+	}
+}
+
+func TestQueryMetrics(t *testing.T) {
+	for _, tc := range []struct {
+		desc  string
+		in    string
+		exp   interface{}
+		error bool
+	}{
+		{
+			desc: "valid time duration",
+			in:   `2s`,
+			exp:  2 * time.Second,
+		},
+		{
+			desc: "empty header",
+			in:   ``,
+			exp:  nil,
+		},
+		{
+			desc: "invalid time duration",
+			in:   `foo`,
+			exp:  nil,
+		},
+	} {
+		t.Run(tc.desc, func(t *testing.T) {
+			req := httptest.NewRequest("GET", "http://testing.com", nil)
+			req.Header.Set(string(QueryQueueTimeHTTPHeader), tc.in)
+
+			w := httptest.NewRecorder()
+			checked := false
+			mware := ExtractQueryMetricsMiddleware().Wrap(http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
+				require.Equal(t, tc.exp, req.Context().Value(QueryQueueTimeHTTPHeader))
 				checked = true
 			}))